+++ release/plugins/utils/index.es6.js
@@ -3,4 +3,5 @@ import './dts/index.es6.js';
 import './env/index.es6.js';
 import './protection/index.es6.js';
 import './swagger/index.es6.js';
+import './ui/index.es6.js';
 import './whitelist/index.es6.js';