+++ release/plugins/utils/index.js
@@ -5,5 +5,6 @@ require('./dts/index.js');
 require('./env/index.js');
 require('./protection/index.js');
 require('./swagger/index.js');
+require('./ui/index.js');
 require('./whitelist/index.js');