+++ release/plugins/utils/swagger/swagger.d.ts
@@ -2,4 +2,5 @@ import { type HandlerPlugin } from 'innet';
 export interface SwaggerProps {
     path?: string;
 }
+/** @deprecated Use <ui> */
 export declare const swagger: HandlerPlugin;